requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.0

# Planned for Step 2 (pin once integrated)
pinecone
google-generativeai
cohere
numpy
scikit-learn
